[BUGFIX] Fix permission checks for access restricted tags
Permission checks for access restricted tags for the policy `readViewnumber` and `resetViews` does currently not work (e.g. no read view count is shown), because the permission check uses `discussion.readViewnumber` instead of `readViewnumber`. For tags, which are not access restricted this is no problem, because the final access check in `Flarum\User\Access\Gate::allows()` will allow `discussion.readViewnumber` in `$actor->hasPermission($ability)`.

In order to resolve the problem, the `discussion.` string has to be removed, since this is appended by `flarum/tags` Access policy checks.